+++ src/game/game.scheduler.ts
@@ -1,9 +1,12 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, Logger } from '@nestjs/common';
 import { Cron, CronExpression } from '@nestjs/schedule';
+import { environment } from '../environment';
 import { GameService } from './game.service';
 
 @Injectable()
 export class GameScheduler {
+  private logger = new Logger('Game Cleaner');
+
   constructor(
     private gameService: GameService,
   ) {
@@ -11,7 +14,10 @@ export class GameScheduler {
 
   @Cron(CronExpression.EVERY_10_MINUTES)
   async deleteEmptyGames(): Promise<void> {
-    const oneHourInMillis = 60 * 60 * 1000;
-    await this.gameService.deleteEmptyGames(oneHourInMillis);
+    const maxAgeMs = environment.cleanup.deleteGameAfterHours * 60 * 60 * 1000;
+    const games = await this.gameService.deleteOldGames(maxAgeMs);
+    if (games.length) {
+      this.logger.warn(`Deleted ${games.length} old games.`);
+    }
   }
 }