+++ src/game/game.service.ts
@@ -57,11 +57,10 @@ export class GameService {
     return deleted;
   }
 
-  async deleteEmptyGames(olderThanMs: number): Promise<Game[]> {
+  async deleteOldGames(olderThanMs: number): Promise<Game[]> {
     const filterDate = new Date(Date.now() - olderThanMs);
     const games = await this.model.find({
       updatedAt: { $lt: filterDate },
-      members: 0,
     }).exec();
     await this.model.deleteMany({ _id: { $in: games.map(g => g._id) } });
     for (const game of games) {