+++ src/group/group.module.ts
@@ -1,8 +1,10 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
+import { environment } from '../environment';
 
 import { EventModule } from '../event/event.module';
 import { GroupController } from './group.controller';
+import { GroupScheduler } from './group.scheduler';
 import { GroupSchema } from './group.schema';
 import { GroupService } from './group.service';
 
@@ -17,7 +19,7 @@ import { GroupService } from './group.service';
     EventModule,
   ],
   controllers: [GroupController],
-  providers: [GroupService],
+  providers: environment.passive ? [GroupService] : [GroupService, GroupScheduler],
   exports: [GroupService],
 })
 export class GroupModule {