+++ src/message/message.dto.ts
@@ -1,4 +1,6 @@
-import { PartialType, PickType } from '@nestjs/swagger';
+import { ApiPropertyOptional, PartialType, PickType } from '@nestjs/swagger';
+import { Transform } from 'class-transformer';
+import { IsDate, IsInt, IsOptional, Max, Min } from 'class-validator';
 import { Message } from './message.schema';
 
 export class CreateMessageDto extends PickType(Message, [
@@ -9,4 +11,28 @@ export class CreateMessageDto extends PickType(Message, [
 export class UpdateMessageDto extends PartialType(PickType(Message, [
   'body',
 ] as const)) {
+}
+
+const MESSAGE_LIMIT = 100;
+
+export class QueryMessagesDto {
+  @ApiPropertyOptional({ description: 'The timestamp before which messages are requested' })
+  @Transform(({ value }) => value ? new Date(value) : undefined)
+  @IsOptional()
+  @IsDate()
+  createdBefore?: Date;
+
+  @ApiPropertyOptional({
+    description: 'The maximum number of results',
+    type: 'integer',
+    minimum: 1,
+    maximum: MESSAGE_LIMIT,
+    default: MESSAGE_LIMIT,
+  })
+  @Transform(({ value }) => value ? +value : undefined)
+  @IsOptional()
+  @IsInt()
+  @Min(1)
+  @Max(MESSAGE_LIMIT)
+  limit?: number;
 }