+++ src/message/message.handler.ts
@@ -1,6 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { OnEvent } from '@nestjs/event-emitter';
-import { MemberResolverService } from '../member-resolver/member-resolver.service';
+import { Game } from '../game/game.schema';
+import { Group } from '../group/group.schema';
+import { MemberResolverService, Namespace } from '../member-resolver/member-resolver.service';
 import { MessageService } from './message.service';
 
 @Injectable()
@@ -11,11 +13,19 @@ export class MessageHandler {
   ) {
   }
 
-  @OnEvent('*.*.deleted')
-  async onDeleted(entity: any): Promise<void> {
-    const resolved = await this.memberResolver.resolveFrom(entity);
-    if (resolved) {
-      await this.messageService.deleteAll(resolved[0], entity._id, resolved[1]);
-    }
+  @OnEvent('games.*.deleted')
+  async onGameDeleted(game: Game): Promise<void> {
+    return this.onDelete(Namespace.games, game);
+  }
+
+  @OnEvent('groups.*.deleted')
+  async onGroupDeleted(group: Group): Promise<void> {
+    return this.onDelete(Namespace.groups, group);
+  }
+
+  private async onDelete(namespace: Namespace, entity: any): Promise<void> {
+    const { _id } = entity;
+    const members = await this.memberResolver.resolve(namespace, _id);
+    await this.messageService.deleteAll(namespace, _id, members);
   }
 }