+++ src/user/user.schema.ts
@@ -1,15 +1,16 @@
 import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
-import { ApiProperty, OmitType } from '@nestjs/swagger';
-import { IsByteLength, IsIn, IsNotEmpty, IsOptional, IsString, Matches, MaxLength } from 'class-validator';
+import { ApiProperty } from '@nestjs/swagger';
+import { IsByteLength, IsIn, IsNotEmpty, IsOptional, IsString, MaxLength } from 'class-validator';
 import { Document } from 'mongoose';
 import { GLOBAL_SCHEMA_OPTIONS, GlobalSchema } from '../util/schema';
+import { IsUrlOrUri } from '../util/url-or-uri.validator';
 
 const MAX_AVATAR_LENGTH = 16 * 1024;
 export const STATUS = ['online', 'offline'] as const;
 export type Status = typeof STATUS[number];
 
-@Schema({ ...GLOBAL_SCHEMA_OPTIONS, timestamps: false })
-export class User extends OmitType(GlobalSchema, ['createdAt', 'updatedAt']) {
+@Schema(GLOBAL_SCHEMA_OPTIONS)
+export class User extends GlobalSchema {
   @Prop({ index: { type: 1, unique: true } })
   @ApiProperty({ minLength: 1, maxLength: 32 })
   @IsString()
@@ -24,7 +25,7 @@ export class User extends OmitType(GlobalSchema, ['createdAt', 'updatedAt']) {
 
   @Prop()
   @IsOptional()
-  @Matches(/^\w+:/, { message: 'avatar must be a valid URI' })
+  @IsUrlOrUri()
   @IsByteLength(0, MAX_AVATAR_LENGTH)
   @ApiProperty({ format: 'url', required: false, maxLength: MAX_AVATAR_LENGTH })
   avatar?: string;
@@ -33,7 +34,7 @@ export class User extends OmitType(GlobalSchema, ['createdAt', 'updatedAt']) {
   passwordHash?: string;
 
   @Prop({ transform: () => undefined })
-  refreshKey?: string;
+  refreshKey?: string | null;
 }
 
 export type UserDocument = User & Document;